Move Rust processor config into lib
