+++ rust/processor/src/lib.rs
@@ -11,8 +11,11 @@
 #[macro_use]
 extern crate diesel;
 
+mod config;
 pub mod models;
 pub mod processors;
 pub mod schema;
 mod utils;
 pub mod worker;
+
+pub use config::IndexerGrpcProcessorConfig;
